Updating Acronyms

+++ catscripts/Language/Acronyms/Acronyms.sh
@@ -1,26 +1,30 @@
 #!/bin/bash
 
-if [ "$DEBUG" == "yes" ];
+KEYWORDS_ACRONYMS="acronym|initialism|What does [A-Z]{2,} stand(|s) for|What do the initials [A-Z]{2,} stand(|s) for"
+
+if [ "$1" == "" ]; #Normal operation
 then
-  printf "Starting Acronyms\n" 
-fi
 
-egrep -i 'acronym|initialism' newpages.txt >> Acronyms.txt
-egrep '(^|\[\[)What does [A-Z]{2,} stand(|s) for($|\]\])' newpages.txt >> Acronyms.txt
-egrep '(^|\[\[)What do the initials [A-Z]{2,} stand(|s) for' newpages.txt >> Acronyms.txt
+  if [ "$DEBUG" == "yes" ];
+  then
+    printf "Starting Acronyms\n" 
+  fi
 
-ACRONYMS=`stat --print=%s Acronyms.txt`
+  ACRONYMS=`egrep -i "$KEYWORDS_ACRONYMS" newpages.txt`
 
-if [ $ACRONYMS -ne 0 ];
-then
-  export CATFILE="Acronyms.txt"
-  export CATNAME="Acronyms"
-  $CATEGORIZE
-fi
+  if [ "$ACRONYMS" != "" ];
+  then
+    printf "$ACRONYMS" > Acronyms.txt
+    export CATFILE="Acronyms.txt"
+    export CATNAME="Acronyms"
+    $CATEGORIZE
+    rm Acronyms.txt
+    unset ACRONYMS
+  fi
 
-rm Acronyms.txt
+  if [ "$DEBUG" == "yes" ];
+  then
+    printf "Finishing Acronyms\n" 
+  fi
 
-if [ "$DEBUG" == "yes" ];
-then
-  printf "Finishing Acronyms\n" 
 fi
